+++ src/FEM/FEMSpaces.jl
@@ -22,8 +22,8 @@ function get_lagrangianQuad_info(
   ξₖ = get_cell_map(Ω)
   Qₕ_cell_point = get_cell_points(Qₕ)
   qₖ = get_data(Qₕ_cell_point)
-  phys_quadp = collect(lazy_map(Gridap.evaluate,ξₖ,qₖ))::Vector{Vector{VectorValue{Dp,Float64}}}
-  refFE_quad = Gridap.ReferenceFE(lagrangianQuad,Float64,FEMInfo.order)
+  phys_quadp = collect(lazy_map(Gridap.evaluate,ξₖ,qₖ))::Vector{Vector{VectorValue{Dp,Float}}}
+  refFE_quad = Gridap.ReferenceFE(lagrangianQuad,Float,FEMInfo.order)
   V₀_quad = TestFESpace(model,refFE_quad,conformity=:L2)
 
   phys_quadp, V₀_quad
@@ -39,8 +39,8 @@ function get_lagrangianQuad_info(
   ξₖ = get_cell_map(Ω)
   Qₕ_cell_point = get_cell_points(Qₕ)
   qₖ = get_data(Qₕ_cell_point)
-  phys_quadp = collect(lazy_map(Gridap.evaluate,ξₖ,qₖ))::Vector{Vector{VectorValue{Dp,Float64}}}
-  refFE_quad = Gridap.ReferenceFE(lagrangianQuad,Float64,FEMInfo.order)
+  phys_quadp = collect(lazy_map(Gridap.evaluate,ξₖ,qₖ))::Vector{Vector{VectorValue{Dp,Float}}}
+  refFE_quad = Gridap.ReferenceFE(lagrangianQuad,Float,FEMInfo.order)
   V₀_quad = TestFESpace(model,refFE_quad,conformity=:L2)
 
   phys_quadp, V₀_quad
@@ -56,7 +56,7 @@ function get_FEMSpace(
   Ω, Γn, Qₕ, dΩ, dΓn, dΓd = get_mod_meas_quad(FEMInfo, model)
 
   labels = set_labels(FEMInfo, model)
-  refFE = Gridap.ReferenceFE(lagrangian, Float64, FEMInfo.order)
+  refFE = Gridap.ReferenceFE(lagrangian, Float, FEMInfo.order)
 
   V₀ = TestFESpace(model, refFE; conformity=:H1,
     dirichlet_tags=FEMInfo.dirichlet_tags, labels=labels)
@@ -83,7 +83,7 @@ function get_FEMSpace(
 
   Ω, Γn, Qₕ, dΩ, dΓn, dΓd = get_mod_meas_quad(FEMInfo, model)
 
-  refFE = Gridap.ReferenceFE(lagrangian, Float64, FEMInfo.order)
+  refFE = Gridap.ReferenceFE(lagrangian, Float, FEMInfo.order)
   labels = set_labels(FEMInfo, model)
   V₀ = TestFESpace(model, refFE; conformity=:H1,
     dirichlet_tags=FEMInfo.dirichlet_tags, labels=labels)
@@ -110,7 +110,7 @@ function get_FEMSpace(
 
   Ω, Γn, Qₕ, dΩ, dΓn, dΓd = get_mod_meas_quad(FEMInfo, model)
 
-  refFEᵤ = Gridap.ReferenceFE(lagrangian, VectorValue{D,Float64}, FEMInfo.order)
+  refFEᵤ = Gridap.ReferenceFE(lagrangian, VectorValue{D,Float}, FEMInfo.order)
   labels = set_labels(FEMInfo, model)
   V₀ = TestFESpace(model, refFEᵤ; conformity=:H1,
     dirichlet_tags=FEMInfo.dirichlet_tags, labels=labels)
@@ -120,7 +120,7 @@ function get_FEMSpace(
   ϕᵤ = get_trial_fe_basis(V)
   Nₛᵘ = length(get_free_dof_ids(V))
 
-  refFEₚ = Gridap.ReferenceFE(lagrangian, Float64, order - 1; space=:P)
+  refFEₚ = Gridap.ReferenceFE(lagrangian, Float, order - 1; space=:P)
   Q₀ = TestFESpace(model, refFEₚ; conformity=:L2, constraint=:zeromean)
   Q = TrialFESpace(Q₀)
   ψᵧ = get_trial_fe_basis(Q₀)
@@ -148,7 +148,7 @@ function get_FEMSpace(
 
   Ω, Γn, Qₕ, dΩ, dΓn, dΓd = get_mod_meas_quad(FEMInfo, model)
 
-  refFEᵤ = Gridap.ReferenceFE(lagrangian, VectorValue{D,Float64}, FEMInfo.order)
+  refFEᵤ = Gridap.ReferenceFE(lagrangian, VectorValue{D,Float}, FEMInfo.order)
   labels = set_labels(FEMInfo, model)
   V₀ = TestFESpace(model, refFEᵤ; conformity=:H1,
     dirichlet_tags=FEMInfo.dirichlet_tags, labels=labels)
@@ -158,7 +158,7 @@ function get_FEMSpace(
   ϕᵤ(t) = get_trial_fe_basis(V(t))
   Nₛᵘ = length(get_free_dof_ids(V₀))
 
-  refFEₚ = Gridap.ReferenceFE(lagrangian, Float64, FEMInfo.order - 1; space=:P)
+  refFEₚ = Gridap.ReferenceFE(lagrangian, Float, FEMInfo.order - 1; space=:P)
   Q₀ = TestFESpace(model, refFEₚ, conformity=:L2, constraint=:zeromean)
   Q = TrialFESpace(Q₀)
   ψᵧ = get_fe_basis(Q₀)
@@ -203,7 +203,7 @@ function get_FEMSpace₀(
   FEMInfo::SteadyInfo,
   model::DiscreteModel)
 
-  get_FEMSpace(problem_id,FEMInfo,model, x->zero(VectorValue(FEMInfo.D, Float64)))
+  get_FEMSpace(problem_id,FEMInfo,model, x->zero(VectorValue(FEMInfo.D, Float)))
 
 end
 
@@ -212,7 +212,7 @@ function get_FEMSpace₀(
   FEMInfo::UnsteadyInfo,
   model::DiscreteModel)
 
-  g₀(x, t::Real) = zero(VectorValue(FEMInfo.D, Float64))
+  g₀(x, t::Real) = zero(VectorValue(FEMInfo.D, Float))
   g₀(t::Real) = x -> g₀(x, t)
   get_FEMSpace(problem_id,FEMInfo,model,g₀)
