+++ src/RB/RBStokes_steady.jl
@@ -472,8 +472,8 @@ function online_phase(
   end
 
   pass_to_pp = Dict("path_μ"=>path_μ, "FEMSpace"=>FEMSpace,
-    "mean_point_err_u"=>Float64.(mean_pointwise_err_u),
-    "mean_point_err_p"=>Float64.(mean_pointwise_err_p))
+    "mean_point_err_u"=>Float.(mean_pointwise_err_u),
+    "mean_point_err_p"=>Float.(mean_pointwise_err_p))
 
   if RBInfo.post_process
     post_process(RBInfo, pass_to_pp)
